+++ kDriveCore/Data/UploadQueue/Servicies/PhotoLibraryUploader+Clean.swift
@@ -19,3 +19,65 @@
 import Foundation
 import Photos
 
+public extension PhotoLibraryUploader {
+    /// Wrapper type to map an "UploadFile" to a "PHAsset"
+    struct PicturesAssets {
+        /// Collection of primary keys of "UploadFile"
+        public let filesPrimaryKeys: [String]
+
+        /// collection of PHAsset
+        public let assets: PHFetchResult<PHAsset>
+    }
+
+    func getPicturesToRemove() -> PicturesAssets? {
+        Log.photoLibraryUploader("getPicturesToRemove")
+        // Check that we have photo sync enabled with the delete option
+        guard let settings, settings.deleteAssetsAfterImport else {
+            Log.photoLibraryUploader("no settings")
+            return nil
+        }
+
+        var toRemoveFileIDs = [String]()
+        var toRemoveAssets = PHFetchResult<PHAsset>()
+        BackgroundRealm.uploads.execute { realm in
+            toRemoveFileIDs = uploadQueue
+                .getUploadedFiles(using: realm)
+                .filter("rawType = %@", UploadFileType.phAsset.rawValue)
+                .map(\.id)
+            toRemoveAssets = PHAsset.fetchAssets(withLocalIdentifiers: toRemoveFileIDs, options: nil)
+        }
+
+        guard toRemoveAssets.count >= Self.removeAssetsCountThreshold,
+              uploadQueue.operationQueue.operationCount == 0 else {
+            return nil
+        }
+
+        return PicturesAssets(filesPrimaryKeys: toRemoveFileIDs, assets: toRemoveAssets)
+    }
+
+    func removePicturesFromPhotoLibrary(_ toRemoveItems: PicturesAssets) {
+        Log.photoLibraryUploader("removePicturesFromPhotoLibrary toRemoveItems:\(toRemoveItems.filesPrimaryKeys.count)")
+        PHPhotoLibrary.shared().performChanges {
+            PHAssetChangeRequest.deleteAssets(toRemoveItems.assets)
+        } completionHandler: { success, _ in
+            guard success else {
+                return
+            }
+
+            BackgroundRealm.uploads.execute { realm in
+                do {
+                    try realm.write {
+                        let filesInContext = realm
+                            .objects(UploadFile.self)
+                            .filter("id IN %@", toRemoveItems.filesPrimaryKeys)
+                            .filter { $0.isInvalidated == false }
+                        realm.delete(filesInContext)
+                    }
+                    Log.photoLibraryUploader("removePicturesFromPhotoLibrary success")
+                } catch {
+                    Log.photoLibraryUploader("removePicturesFromPhotoLibrary error:\(error)", level: .error)
+                }
+            }
+        }
+    }
+}
